requirements.txt
cached_property==2.0.1
click==8.1.7
fastapi==0.122.1
httpx==0.28.1
Jinja2==3.1.6
nodejs==0.1.1
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python_jose==3.5.0
regex==2024.11.6
rich==14.2.0
strenum==0.4.15
tomlkit==0.13.3
typing_extensions==4.15.0
